fix: unbreak the 32-bit build, make RRF order-independent, catch typed nils
PR #1 review round. Six threads were open; five were valid and are fixed, the
sixth is answered here and left for the author to close.

The full-index guard added last commit does not compile on a 32-bit target.
len returns int, so an untyped math.MaxUint32 on the other side of the
operator becomes an int, and 4294967295 does not fit in one: GOARCH=386 fails
the whole package with "overflows int". Reproduced, then fixed by widening the
count to uint64 before comparing, where the test is simply never true — which
is the right answer, since an int that narrow cannot reach the limit. 386, arm,
amd64 and arm64 all build now. A guard against a bound no corpus will reach
had made the engine unavailable on an entire class of target.

fusion.Fuse accumulated stream by stream, so the order a document's
contributions arrived in was the caller's scorer order, and float addition is
not associative. Two documents holding the same multiset of ranks — 1,2,7
against 7,1,2 — got mathematically equal totals that differ in the last bit,
so permuting the scorer slice picked a winner instead of letting TopK's DocID
tiebreak settle it. Confirmed by search before fixing: ranks 1,2,7 are the
smallest such triple, and their six permutations produce two distinct float64
values. Sweeping rank-major instead of stream-major makes every document
accumulate in ascending rank order whatever stream supplied each hit, so equal
multisets sum to identical bits; repeats of one rank across streams add the
same value and commute. Same total work, one reciprocal per rank rather than
per candidate, and the new test fails against the old loop.

The same loop restructure fixes the map hint. Sizing it k*len(streams) treated
a caller-supplied k as a bound on real candidates — cmd/weft takes k straight
from a flag, so one large -k reserved space proportional to a number of
documents that does not exist. Summed stream lengths are exact and cannot
overflow, since those elements are already in memory.

Search now rejects a typed-nil scorer, not just a nil interface. `var s
*text.Scorer` left unassigned is not == nil, because the interface carries a
type descriptor, and that is the form runtime configuration actually produces.
Last round documented the hole rather than closing it, on the grounds that
reflect was heavy for a leaf package. That reasoning does not survive contact
with index.go's own promise above its sentinels — library code here never
panics — and a guard that rejects the nil a caller did not write while
panicking on the one they did is worse than either alternative. One reflect
call per scorer per query, against three full corpus scans. Only Pointer is
checked: a nil map, slice or func receiver does not fault on a method call the
way a dereferenced nil pointer does, and no scorer in the tree is any of those.

The API golden recorded exported fields only, so Document gaining its first
unexported field would have passed. Go forbids an unkeyed composite literal
from another package once a struct holds one, so that change breaks every
external engine.Document{k, t, v, l, ts} while changing no name and no
exported type. Verified both halves by mutation: an external unkeyed literal
compiles today and stops compiling with the field added, and the golden now
reports it. Recorded as one marker per struct rather than a line per field —
the second unexported field takes away nothing the first had not, their names
are unreachable and their order is not promised, so recording each would fail
the assertion on every internal Index field and demand an engine cost that
does not exist. That is the fourth correction to this assertion and the second
in the direction of recording less; FINDINGS section 1 has the table.

Not fixed, and the reason is in the code: Search checks the context before the
fuser but deliberately not after it. Every context check in this tree buys
something by declining work about to be wasted. A check on the way out
declines nothing — the fusion is paid for and the candidates are correct — so
it would trade a complete answer for an error, on a caller who is free to drop
the result themselves. Search promises that a cancelled query does not do
fusion, not that a cancelled query never returns one.

Every fix carries a test that fails without it, verified by mutation in both
directions for the golden. Line counts: 1,185 implementation, 2,457 test. make
arch passes 7 assertions, recency holds at 99 of its 100-line budget, the demo
output is unchanged, and the suite is green under -race.

+++ pkg/engine/index.go
@@ -156,7 +156,13 @@ func (ix *Index) Add(d Document) (DocID, error) {
 	// afterwards addresses document 0 instead. Doc and DocLen already compare
 	// their bounds in uint64 to avoid the mirror image of this; refusing the Add
 	// is the same choice on the write side, and Add already returns an error.
-	if len(ix.docs) > math.MaxUint32 {
+	//
+	// Widened to uint64 before comparing, not after. len returns int, so an
+	// untyped MaxUint32 on the other side of the operator becomes an int, which
+	// does not fit on a 32-bit target: the package stops compiling entirely
+	// under GOARCH=386. Widened, the comparison is simply never true there,
+	// which is the right answer — an int that narrow cannot reach the limit.
+	if uint64(len(ix.docs)) > math.MaxUint32 {
 		return 0, fmt.Errorf("add %q: index is full at %d documents", d.Key, uint64(math.MaxUint32)+1)
 	}
 	id := DocID(len(ix.docs))
